feat: quiz com pontos de fé, migração int→double, gráfico dinâmico
- Quiz aplica faith points ao responder (apply_faith_penalty no _saveAttempt)
- Desbloqueio de quiz não dá mais pontos (só ao responder)
- Admin: campo "Ao errar" auto-converte para negativo
- Migração int→double em todos os campos de fé (~17 arquivos)
- Gráfico de fé com escala automática e valores sempre visíveis
- Fix snackbar stacking (clearSnackBars antes de showSnackBar)
- Fix contagem de perguntas do quiz (era hardcoded "5 perguntas")
- Merge admin_quiz_tab + admin_quizzes_tab em arquivo único
- Remoção de build artifacts do assets/

+++ lib/features/diary/utils/faith_penalty.dart
@@ -1,8 +1,8 @@
 import '../models/scoring_word.dart';
 
-int calculateContentScore(String content, List<ScoringWord> words) {
+double calculateContentScore(String content, List<ScoringWord> words) {
   final lower = content.toLowerCase();
-  int total = 0;
+  double total = 0;
 
   for (final sw in words) {
     if (lower.contains(sw.word.toLowerCase())) {
